test(soft-rw): reuse existing test module
